requirements_model.txt
numpy==1.19.0
pandas==1.0.5
pyyaml==5.3.1
tqdm==4.48.0
cachetools==4.1.1
cffi==1.14.1
google-api-core==1.22.0
google-auth==1.20.0
google-cloud-core==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
rsa==4.6
cloudml-hypertune==0.1.0.dev6